DCAC-241: Log outcome of update propagation, refactor rest template config.
