+++ src/main/java/uk/gov/companieshouse/itemgroupworkflowapi/config/ApplicationConfiguration.java
@@ -1,17 +1,22 @@
 package uk.gov.companieshouse.itemgroupworkflowapi.config;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.datatype.jsonp.JSONPModule;
+import java.util.ArrayList;
+import java.util.List;
 import org.apache.http.impl.client.HttpClients;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
@@ -34,7 +39,6 @@ public class ApplicationConfiguration {
                 .build();
     }
 
-    // TODO DCAC-241 Remove these REST templates
     @Bean
     public RestTemplateBuilder getRestTemplateBuilder() {
         return new RestTemplateBuilder();
@@ -44,7 +48,23 @@ public class ApplicationConfiguration {
     public RestTemplate restTemplate(RestTemplateBuilder builder) {
         final var httpClient = HttpClients.custom().disableRedirectHandling().build();
         final var requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
-        return builder.requestFactory(() -> requestFactory).build();
+        return builder.additionalMessageConverters(getJsonMessageConverters())
+                      .requestFactory(() -> requestFactory)
+                      .build();
+    }
+
+    private List<HttpMessageConverter<?>> getJsonMessageConverters() {
+        final List<HttpMessageConverter<?>> converters = new ArrayList<>();
+        final var converter = new MappingJackson2HttpMessageConverter();
+        converter.setObjectMapper( createObjectMapper());
+        converters.add(converter);
+        return converters;
+    }
+
+    private ObjectMapper createObjectMapper() {
+        final var objectMapper = new ObjectMapper();
+        objectMapper.configure(MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES, true);
+        return objectMapper;
     }
 
 }
